+++ src/common/nest-string.js
@@ -1,0 +1,11 @@
+function getNestString(vParent, vChild)
+{
+	var stringRes = [vParent, vChild].join(".");
+	return stringRes;
+}
+
+
+module.exports =
+{
+	get: getNestString
+};