+++ changelog.md
@@ -1,15 +1,54 @@
 # Changelog
 
-**./src/input/read-line-data.js**
-* readDataFile
-	* Removed spaces from event function argument lists.
-* readCurrentLine
-	* Declared new variable 'readLineNum' - Shortcut for current line number.
+**./src/common/nest-string.js**
+* New file - Used to write nest string for parent and child option properties.
+	* "parent.child"
+* Derived from: `../options-validation/list-entry-props.js`
 
 ---
 
-**./src/input/read-first-names.js**
-* Split the 'checkDataLineLength' call into its own function 'callDataLengthCheck'
-* parseRowString
-	* Renamed the 'validNameLength' variable to 'validName'
-	* Renamed the 'currentLineNumber' parameter to 'fileLineNum'
+**./src/options-validation/list-entry-props.js**
+* Moved the 'getNestString' function into: '../common/nest-string.js'
+	* This file has been added as a requirement.
+* Replaced 'getNestString' calls with 'nestString.get'
+* Renamed 'nestString' variable to 'propString' in functions:
+	* handleChancePercentage
+	* handleRangeNumber
+	* handleTrueFalse
+
+---
+
+**./src/options-validation/keyword-props.js**
+* Added requirement for: '../common/nest-string.js'
+* handleChanceProperty
+	* Renamed the 'nestString' variable to 'propString'
+* handleKeywordProperty
+	* Renamed the 'nestString' variable to 'propString'
+	* 'propString' is now assigned using 'nestString.get'
+
+---
+
+**./src/options-validation/age-props.js - handleAgeNumber**
+* Renamed 'nestString' variable to 'propString'
+
+---
+
+**./src/options-validation/basic-props.js - handleGenderPercentage**
+* Renamed 'nestString' variable to 'propString'
+
+---
+
+**./src/options-validation/pet-props.js - handleCountNumber**
+* Renamed 'nestString' variable to 'propString'
+
+---
+
+**./src/options-validation/prev-experience-props.js - handleRangeNumber**
+* Renamed 'nestString' variable to 'propString'
+
+---
+
+**./src/options-validation/range-props.js**
+* Renamed 'nestString' variable to 'propString' in functions:
+	* handleTimeNumber
+	* handleViewNumber